requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.5
librosa>=0.10.0
numpy>=1.24.0
soundfile>=0.12.0
pydub>=0.25.1
scikit-learn>=1.2.0
joblib>=1.2.0
openai-whisper>=20230314
torch>=2.0.0
